Create post.py
This runs through all posts and it actually breaks out if there is an error. It keeps track of completed URLS in a log file. To continue where you left off, just tail the end of the log file and go from the last successful post ID. A lot more can be done with this